# Fix and optimize `bin1d_vec` + extend units tests + fix imprecise bin edge creation