--- a/tests/test_regions.py
+++ b/tests/test_regions.py
@@ -86,3 +86,72 @@ class TestNZRegion(unittest.TestCase):
         r = nz_csep_region()
         # they dont have to be in the same order, but they need
         numpy.testing.assert_array_equal(r.midpoints().sort(), self.from_dat.sort())
+
+class TestBin2d(unittest.TestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        super(TestBin2d, cls).setUpClass()
+
+        # (loading those is the bottleneck of this test case)
+        cls.regions = [
+            italy_csep_region(),
+            california_relm_region(),
+            nz_csep_region()
+        ]
+
+    def test_bin2d_regions_origins(self):
+        """every origin must be inside its own bin
+        """
+
+        for region in self.regions:
+            origins = region.origins()
+            self._test_bin2d_region_loop(region, origins)
+            self._test_bin2d_region_vect(region, origins)
+            self._test_bin2d_region_vect(region, origins.astype(numpy.float32))
+
+    def test_bin2d_regions_midpoints(self):
+        """every midpoint must be inside its own bin
+        """
+
+        for region in self.regions:
+            midpoints = region.midpoints()
+            self._test_bin2d_region_loop(region, midpoints)
+            self._test_bin2d_region_vect(region, midpoints)
+            self._test_bin2d_region_vect(region, midpoints.astype(numpy.float32))
+
+    def test_bin2d_regions_endcorner(self):
+        """every corner point (~opposite end of the origin) must be inside its own bin
+        """
+
+        for region in self.regions:
+            frac = 0.9999999999
+            endcorners = region.origins() + frac*region.dh
+            self._test_bin2d_region_loop(region, endcorners)
+            self._test_bin2d_region_vect(region, endcorners)
+            frac = 0.999  # decrease frac for float32 due to its lower resolution
+            endcorners = region.origins() + frac*region.dh
+            self._test_bin2d_region_vect(region, endcorners.astype(numpy.float32))
+
+    def _test_bin2d_region_loop(self, region, coords):
+        """(slow) loop over origins; each time, calls bin1d_vec for lat & lon scalars
+        """
+
+        for i, origin in enumerate(coords):
+            idx = region.get_index_of(
+                origin[0],
+                origin[1],
+            )
+            self.assertEqual(i, idx)
+
+    def _test_bin2d_region_vect(self, region, coords):
+        """call bin1d_vec once for all lat origins & all lon origins
+
+        Besides, also tests if vectors with ndim=2 are consumed properly
+        (returns 2nd-order/nested list ([[...]]).
+        """
+
+        lons, lats = numpy.split(coords.T, 2)  # both have ndim=2!
+        test = region.get_index_of(lons, lats).tolist()  # nested list ([[...]])
+        expected = [numpy.arange(len(region.origins())).tolist()]  # embed in another list
+        self.assertListEqual(test, expected)