⚠ Caution, longer treatise ahead! 🤷‍♂️

### 1. Solving numerical issues (#255)

I [previously proposed a preliminary fix](https://github.com/SCECcode/pycsep/issues/255#issuecomment-2548687127) based on rounding the estimated bin size `h`. But there's a more elegant [fix](https://github.com/SCECcode/pycsep/commit/e12f6a0):

```diff
-    h_tol = numpy.abs(h) * numpy.finfo(numpy.float64).eps
+    h_tol = a0_tol
```
in [this line](https://github.com/SCECcode/pycsep/blob/2feb9148/csep/utils/calc.py#L83).

This change is sufficient because the prior floating-point operation `h = bins[1] - bins[0]` in the [previous line](https://github.com/SCECcode/pycsep/blob/2feb9148/csep/utils/calc.py#L83) carries the unit [roundoff error](https://en.wikipedia.org/wiki/Round-off_error) of the involved numbers (and may introduce a further roundoff error). To account for it, the tolerance must be based on machine `eps`ilon and the _involved_ numbers (not the resulting number). In our case, the two involved numbers in calculating `h` are of the same order of magnitude, and one of it is `a0` (i.e., the first bin edge `bin[0] == min(bins)`), for which we already calculate the tolerance (`a0_tol`). So we can use `a0_tol` also as tolerance for `h`.

Here's an illustration:

```
       bin1:   31.50000000000000000 (a0)
       bin2:   31.60000000000000142

          h:    0.10000000000000142 (bin2 - a0)
imprecision:    0.00000000000000142 (|h - 0.1|)

        eps:    0.00000000000000022 (FYI)
    |h|*eps:    0.00000000000000002 (previous h_tol)

   |a0|*eps:    0.00000000000000699 (new h_tol)
```

To account for the floating-point/roundoff errors, `h_tol` must be larger than the `imprecision` of `h`.
Only the new `h_tol` estimation guarantees that by reusing `a0_tol = abs(a0) * eps`.

> FYI: accounting for `h_tol` would not be necessary if the actual bin size `h` is accurately specified (e.g., `0.1`), i.e., not based on a prior operation. `CartesianGrid` already stores it as meta-data `.dh`, which is _exactly_ `0.1` for all currently implemented regions. `bin1d_vec` _could_ be adapted to receive this bin size as an additional argument. But `bin1d_vec` is also [used for binning magnitudes](https://github.com/search?q=repo%3ASCECcode%2Fpycsep+bin1d_vec%28&type=code), for which the bin size is not explicitly stored in the `region`. Of course we could change that and pass it to `bin1d_vec`, but I believe it's not necessary (for now).


### 2. Further enhancements

#### 2.1 to `bin1d_vec`

Working with and looking at `bin1d_vec` way too often, I noticed that it can be simplified, made more robust/versatile _and_ (slightly) speed up by addressing various aspects. In the following, I list each incremental enhancement and measure its performance; I committed them separately for convenience and providing related notes in their description.

To quantify the performance speed-up, I timed the self-check over all origins for California, but – to avoid the overhead in `get_index_of` – only for one coordinate component by directly calling `bin1d_vec` in two ways:
1. looped (7682 individual call): `%timeit -r 300 -n 1 for origin in region.origins(): bin1d_vec(origin[0], region.xs)`
2. vectorized (vector with 7682 values), which matters much more in practice: `%timeit -r 6000 -n 100 bin1d_vec(origins_lon, region.xs)`

| Enhancement                                             | **vectorized** |  looped |
|:--------------------------------------------------------|---------------:|--------:|
| [Only the bare fix](https://github.com/SCECcode/pycsep/commit/e12f6a0)                                       |    **23.8 µs** | 36.3 ms |
| [1. Reorganize + make tolerance depend on _dtype_](https://github.com/SCECcode/pycsep/commit/79c08dd)        |    **24.1 µs** | 37.7 ms |
| [2. Replace `.min(bins)` with `bins[0]`](https://github.com/SCECcode/pycsep/commit/ec74058)                  |    **22.2 µs** | 26.3 ms |
| [3. Use `.asarray()` instead of `.array()`](https://github.com/SCECcode/pycsep/commit/22aac12)               |    **20.7 µs** | 24.8 ms |
| [4. Avoid try/except by always operating on arrays](https://github.com/SCECcode/pycsep/commit/1fac3b)       |    **20.8 µs** | 31.4 ms |

<br>

> Notes about `bin1d_vec` (for those interested):
> * As mentioned in `bin1d_vec`'s docstring, ["`bins [...] must be monotonically increasing`"](https://github.com/SCECcode/pycsep/blob/2feb9148/csep/utils/calc.py#L63) because the implemented `idx` calculation assumes/requires sorted bins (for this reason, there is no need for `.min()`). This is not checked, presumably because _pyCSEP_ itself never passes unsorted bins to this method (they always originate from ranges). Nevertheless, I added a check at the top, but commented it out; we may activate it in the future if there is need - for now, it only serves as an additional hint.
> * for passing the unit tests mentioned below, the necessary tolerances to add in the `idx` calculation depend on the _dtype_ of `p`:
>     * _float64_ doesn't need adding `a0_tol` nor `p_tol`;
>     * _float32_ doesn't need adding `a0_tol`, but `p_tol` (since `p` is _float32_);
>     * --> So `a0_tol` could be omitted, but to avoid unforeseeable future oddities, let's keep it anyway (and because it theoretically makes sense adding it).
> * the grouping and order of terms in the `idx` calculation influence the roundoff error, see [_What Every Computer Scientist Should Know About Floating-Point Arithmetic_](https://docs.oracle.com/cd/E19957-01/806-3568/ncg_goldberg.html). Illustration:
> ```
>  40.40000000000000568 (p)
>  31.50000000000000000 (a0)
>   0.00000000000003728 (p_tol)
>   0.00000000000003679 (a0_tol)
>
>   8.90000000000000568 (p - a0)
>   0.00000000000007407 (p_tol + a0_tol)
>
>   8.90000000000007674  <--  p + (p_tol + a0_tol) - a0  [ORIGINAL]
>   8.90000000000007674  <--  p + p_tol + a0_tol - a0
>   8.90000000000008029  <--  p - a0 + (p_tol + a0_tol)
>   8.90000000000008029  <--  p - a0 + p_tol + a0_tol    [NEW]
>   8.90000000000007851  <--  p + p_tol - a0 + a0_tol
> ```
> > Only the 3rd and 4th option is closest to (and occasionally exactly) the actual sum (and is largest). In the first enhancement step, I used the 4th option. 

#### 2.2 other changes

* in `region.CartesianGrid2D.get_index_of()`, I also [removed `numpy.array()`](https://github.com/SCECcode/pycsep/commit/9c14ee8) (redundant with first lines in `bin1d_vec`; should slightly speed up spatial binning)
* [drop any `tol=0.00001` from `bin1d_vec` calls or functions that call it](https://github.com/SCECcode/pycsep/commit/0b95738) (only used when binning magnitudes), even in the unit tests

<br>

> Notes (for those interested):
> * specifying `tol=0.00001` was not even necessary with the revised tolerance section by [commit 949d342c](https://github.com/SCECcode/pycsep/commit/949d342c) (albeit most of the `tol=0.00001` were introduced by this commit);
>    * Why? because magnitude bins are "cleanly" created with `np.arange()` in [`csep.core.regions.magnitude_bins`](https://docs.cseptesting.org/_modules/csep/core/regions.html#magnitude_bins);
>    * their redundancy is confirmed by passing the extended unit test `test_scalar_inside` (see below) without `tol=0.00001` using the unfixed `bin1d_vec`;
>    * yet, I left the optional `tol=None` argument in `bin1d_vec` and the functions that call it – just in case there is a real need to override the tolerance.


### 3. [Extend unit tests](https://github.com/SCECcode/pycsep/commit/92e33f0470da28edea8371e4687c954d5d8278b1)

* in `test_calc.TestBin1d`:
    * add `test_bin1d_single_bin2`, which inputs a single bin _and_ a single point;
    * split off a new `test_scalar_inside` from `test_scalar_outside`, which checks for (magnitude) bin edges `5.95, 6.05, ..., 8.95`:
        * _all_ corner cases (*.*5)
        * _all_ center bins (*.*0)
        * _all_ end of bins (*.*99999999)
        * one large value (10)
* add separate class `TestBin2d` in `test_region`, which makes more realistic unit tests by extending the self-check mentioned in #255 and performing it...
    * for three regions: Italy, California, NZ (New-Zealand);
    * for all origins, mid-points, and end-corners (in the bin at the opposite side of origin);
    * for double precision (_float64_ / _f8_) and single precision (_float32_ / _f4_) of the _points_ (not bins);
    * as loop (over each point individually) and single vector (all points at once).
    * ==> 36 unit test combinations
    * (albeit targeting `bin1d_vec`, it also unit-tests region's [`CartesianGrid2D.get_index_of`](https://github.com/SCECcode/pycsep/blob/2feb9148/csep/core/regions.py#L618) and by extension [`GriddedForecast.get_index_of`](https://docs.cseptesting.org/reference/generated/csep.core.forecasts.GriddedForecast.get_index_of.html))

<br>

> Notes (for those interested):
> * the _float32_-based unit tests only pass after the first enhancement in the previous section;
>     * (_float32_ is not a use case when using pyCSEP's implemented [readers](https://github.com/SCECcode/pycsep/blob/main/csep/utils/readers.py) [normally](https://docs.cseptesting.org/concepts/catalogs.html#id7) (read values will eventually be converted to `float64` [here](https://github.com/SCECcode/pycsep/blob/2feb9148/csep/core/catalogs.py#L284) by using [these specified dtypes](https://github.com/SCECcode/pycsep/blob/2feb9148/csep/core/catalogs.py#L883)), but it could happen with a custom loader function and the intention to save some memory - note that the [reader module](https://github.com/SCECcode/pycsep/blob/main/csep/utils/readers.py) still reveals some (incorrect) mentions of `float32`/`f4` in the docstrings). 
> * the unit tests for the NZ region only pass after addressing another issue in the next section.


### 4. Another issue: imprecise bin edge creation

For the NZ region, the end-corner-based unit test (`test_bin2d_regions_endcorner`) failed; inspecting the problematic corner point for the Longitude yields:

```
      origin: 167.79999999999998
           p: 167.89998999999997
          a0: 165.69999
           h: 0.09999999999999432
      p - a0: 2.19999999999996
```
It took me a while to spot the culprit; but this helped:
```
region.xs: [165.69999, 165.79999, 165.89999, 165.99999, 166.09999, 166.19999, 166.29999, 166.39999, ...]
dh: 0.1
```

Oops, they are not rounded to the first decimal digit despite `region.dh == 0.1`!

> FYI: The chain leading to failed test:
>  1. improperly rounded `.xs` (`.*9999`)
>  2. ... causes a likewise imprecise `a0`
>  3. ... which causes `(p - a0 + p_tol + a0_tol) / (h - h_tol)` to be just above `22` (`22.000099999901586`), instead of being just below (`21.99999999990184`)
>  4. ... which eventually nudges `idx` into the next bin

Apparently, the underlying [`csep.utils.calc.cleaner_range()`](https://github.com/SCECcode/pycsep/blob/2feb9148/csep/utils/calc.py#L204) is the culprit. It turns out that the particularly chosen `const = 100000` leads to weird numerical imprecision (it doesn't happen if it were 10'000 or 1'000'000):
```
165.7 * 100000  = 16569999.999999998 <-- the chosen `const`
165.7 * 10000   = 1657000.0
165.7 * 1000000 = 165700000.0
```
Then `np.floor(const * start)` operation yields `16569999.0` instead of `1657000.0`.

[Numpy's user guide](https://numpy.org/doc/stable/user/how-to-partition.html) suggest:
> _Use `numpy.linspace` if you want the endpoint to be included in the result, or if you are using a **non-integer step size**._

I tried it, but some bin edges still contained `.*99999...`.

Instead, I [modified `cleaner_range()` only in some other aspects](https://github.com/SCECcode/pycsep/commit/a836927). The crucial modification is using `round` instead of `floor` – very related to the [problematics that lead me to this PR](https://github.com/SCECcode/pycsep/issues/255#issuecomment-2548687127). The other change – replacing the hard-coded `const` with a flexible `scale` parameter – is supposed to account for the number of decimal digits of bin edges _and_ stepping (`const = 100000` would lead to imprecise bin edges if `h < 0.000'01` or if the bin edges start at `*.000'001` – maybe someone needs that in the future 😉).

> Note (for those interested):
> This sub-issue is apparently caused by using `numpy.loadtxt()` inside `nz_csep_region()` instead of [`parse_csep_template()`](https://github.com/SCECcode/pycsep/blob/2feb9148/csep/core/regions.py#L334) as used for reading the California and Italy CSEP region, which uses `float()` from the Python standard library. So this additional fix makes bin creation more robust to input produced with different reader functions.

In the [same commit](https://github.com/SCECcode/pycsep/commit/a836927), I also simplified `core.regions.magnitude_bins()` to just call (this improved) `cleaner_range()`; it was essentially a copy of it.


### 5. Implication on existing test results

As mentioned in my [original comment of issue #255](https://github.com/SCECcode/pycsep/issues/255), test results produced after this PR will become irreproducible with past results **_only if_** the test catalog contains one or more events whose coordinates align with the region's spatial bin edges, e.g., a coarse single-decimal-digit coordinate like 42.3° and a gridding of 0.1°. Otherwise, this issue and PR is irrelevant.

But even if a catalog _does_ contain such events, I don't expect test results would change significantly – perhaps only if _all_ events have coarse locations. Be reminded that I spotted this whole issue only due to a difference at the 3-5th decimal digit for IGPE compared to an independent binning implementation; the test catalog contained five of such events (262 events in total).

I still wanted to assess if this PR leads to some irreproducibility, so I ran our [first reproducibility package](https://github.com/wsavran/pycsep_esrl_reproducibility) using the most recent pyCSEP state (v0.6.3 + 22 commits) – once without and once with all changes in this PR. Eventually, they were exactly the same (up to the last [15th] decimal digit).[^1] Apparently, it doesn't involve events with coarse locations (this is a good thing!).

Apart that, I didn't do other comparisons; feel free to suggest me some (they should contain events with coarse locations and/or involve the NZ region).

---

Closes #255


[^1]: Btw: they were not exactly the same as the [expected output](https://github.com/wsavran/pycsep_esrl_reproducibility/tree/main/expected_output/results) due to some occasional mismatches at the last (15th) decimal digit. These slight mismatches are likely due to using a different platform/OS and/or a different combination of packages (I installed the most up-to-date versions of packages specified in pyCSEP's `requirements.yml` + adjusted `numpy==1.22.4` and `matplotlib==3.5.2` to get a compatible environment for the newest pyCSEP version).