@mjw98 2025-02-06
Wow, Marcus! Thank you!
________________________________
From: Marcus Herrmann ***@***.***>
Sent: 06 February 2025 15:30
To: SCECcode/pycsep ***@***.***>
Cc: Subscribed ***@***.***>
Subject: Re: [SCECcode/pycsep] Fix spatial binning precision + extend units tests + fix imprecise bin edge creation (PR #270)


Bonus: using decimal package

For the sake of curiosity, I also considered the decimal package as suggested by Bill in the original issue<https://github.com/SCECcode/pycsep/issues/255#issuecomment-2515098482> (actually way before this PR and the aha-moment in my comment<https://github.com/SCECcode/pycsep/issues/255#issuecomment-2548687127>).
Here necessary changes:

-    bins = numpy.array(bins)
-    p = numpy.array(p)
+    def array2decimal(x):
+        x = x.squeeze()  # omit superfluous dims
+        if x.ndim:
+            return np.array(
+                list(
+                    map(Decimal,
+                        numpy.array2string(
+                            x
+                            max_line_width=np.inf, threshold=np.inf
+                        )[1:-1].split())
+                    )
+            )
+        else:
+            return np.array(Decimal(numpy.array2string(x)))
+
+
+    p = array2decimal(numpy.array(p))
+    bins = array2decimal(numpy.array(bins))

and

-    # Deal with border cases
-    a0_tol = numpy.abs(a0) * numpy.finfo(numpy.float64).eps
-    h_tol = numpy.abs(h) * numpy.finfo(numpy.float64).eps
-    p_tol = numpy.abs(p) * numpy.finfo(numpy.float64).eps
-
-    # absolute tolerance
-    if tol is None:
-        idx = numpy.floor((p + (p_tol + a0_tol) - a0) / (h - h_tol))
-    else:
-        idx = numpy.floor((p + (tol + a0_tol) - a0) / (h - h_tol))
-    if h < 0:
-        raise ValueError("grid spacing must be positive and monotonically increasing.")
-    # account for floating point uncertainties by considering extreme case
+    idx = (p - a0) / h

But the performance is abysmal due to mapping via strings (that's how the decimals package works to circumvent numerical precision):

Enhancement     vectorized      looped
decimal 17.1 ms 2.25 s

So 320x / 50x slower than the current implementation.

Also, it doesn't pass some unit tests.

Additionally, array2string<https://numpy.org/doc/stable/reference/generated/numpy.array2string.html> must be well configured to avoid creating any weird strings from floats due to numerical precision (alternatively, one could represent the region's origins/bins directly as Decimals on initializations, but that's all not necessary since we solved the numerical issues above).

—
Reply to this email directly, view it on GitHub<https://github.com/SCECcode/pycsep/pull/270#issuecomment-2640148517>, or unsubscribe<https://github.com/notifications/unsubscribe-auth/AD6OISNYUGUFADKWGCL4VD32ON5X3AVCNFSM6AAAAABWTZ63C2VHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDMNBQGE2DQNJRG4>.
You are receiving this because you are subscribed to this thread.Message ID: ***@***.***>