@wsavran 2025-02-11
@mherrmann3 The new solution looks much simpler and more optimal than rounding in the issue. Nice Work! I'm encouraged by your tests with the reproducibility package. I'm looking into why some of the builds are failing for different OS versions right now. It's failing due to an error in the `comcat` package but I haven't tracked down the reason for the `vcr` package throwing errors in some version but not others.