@wsavran 2025-02-11
it looks like it might be related to the pinned vcrpy installation we are using in the workflows/build-test.yml file 
```pip install vcrpy==4.3.1 pytest pytest-cov``` 

there was a warning in the release notes on vcrpy that this version might not play nicely with urllib3. the last successful builds used `urllib3==2.2.3` and the failing builds in this pr are using `urllib3==2.3.0`. 

it looks like we have some options here:
1) remove the pinned version from `.github/workflows/build-test.yml`
2) create a `requirements-test.yml` that includes sphinx and vcrpy (to mimic the `requirements-dev.txt`) and use this in the workflow
3) add vcrpy, pytest, and pytest-cov to the `create-args` directive under `jobs >> steps` in `.github/workflows/build-test.yml`.

@mherrmann3 do you have any suggestions or thoughts on this?