@mherrmann3 2025-02-13
Hey Bill, I'm not experienced with github's workflows, but I'd prefer the simplest option, number 1, in your list.

I don't see any reason for the pinning. It seems that @pabloitu simply pinned `vcrpy` to [`4.3.1`](https://github.com/kevin1024/vcrpy/releases/tag/v4.3.1) as it was the most recent version [back at that time](https://github.com/SCECcode/pycsep/commit/cb54bd04ca1721ac4a582e34b0d791e4d1109bbf) (and then [re-used this part also for `build_pip`](https://github.com/SCECcode/pycsep/commit/c887782a682cbed09106c49319e3b33a207cee66)).

So we could unpin it twice in `.github/workflows/build-test.yml`. You go ahead?